# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IRMS 后端 API（Gin + Gorm + Swagger）",
        "title": "IRMS API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/hosts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hosts"
                ],
                "summary": "列出主机",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (\u003e=1); invalid -\u003e 400",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1~100); invalid -\u003e 400",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword (name/hostname like)",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "type": "string",
                        "description": "Status; invalid -\u003e 400",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "physical",
                            "vm",
                            "cloud_instance",
                            "other"
                        ],
                        "type": "string",
                        "description": "Provider kind; invalid -\u003e 400",
                        "name": "provider_kind",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Location ID; invalid -\u003e 400",
                        "name": "location_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Environment ID; invalid -\u003e 400",
                        "name": "environment_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.HostListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hosts"
                ],
                "summary": "创建主机",
                "parameters": [
                    {
                        "description": "HostCreateRequest (environment_ids required, [] allowed)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.HostCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.IDResponse"
                        }
                    }
                }
            }
        },
        "/hosts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hosts"
                ],
                "summary": "获取主机详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.HostGetResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hosts"
                ],
                "summary": "更新主机",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "HostUpdateRequest (environment_ids required, [] allowed)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.HostUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hosts"
                ],
                "summary": "删除主机",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Host ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.EmptyData": {
            "type": "object"
        },
        "dtoresp.EmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.EmptyData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.HostGetResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/vo.HostVO"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.HostListData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.HostVO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dtoresp.Pagination"
                }
            }
        },
        "dtoresp.HostListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.HostListData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.IDData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtoresp.IDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.IDData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "request.HostCreateRequest": {
            "type": "object",
            "required": [
                "environment_ids",
                "hostname",
                "name",
                "primary_address",
                "provider_kind"
            ],
            "properties": {
                "cloud_instance_id": {
                    "type": "string"
                },
                "cloud_vendor": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "environment_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hostname": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "os_type": {
                    "type": "string"
                },
                "primary_address": {
                    "type": "string"
                },
                "provider_kind": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "request.HostUpdateRequest": {
            "type": "object",
            "required": [
                "environment_ids",
                "hostname",
                "name",
                "primary_address",
                "provider_kind"
            ],
            "properties": {
                "cloud_instance_id": {
                    "type": "string"
                },
                "cloud_vendor": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "environment_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hostname": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "os_type": {
                    "type": "string"
                },
                "primary_address": {
                    "type": "string"
                },
                "provider_kind": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "vo.HostVO": {
            "type": "object",
            "properties": {
                "cloud_instance_id": {
                    "type": "string"
                },
                "cloud_vendor": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "environment_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "environment_source": {
                    "type": "string"
                },
                "environments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inherited_environment_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "location": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "os_type": {
                    "type": "string"
                },
                "own_environment_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "primary_address": {
                    "type": "string"
                },
                "provider_kind": {
                    "type": "string"
                },
                "self_environments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}